Login access behaviour
Also added navbar + profile components to be worked on in other issues
