+++ src/App.js
@@ -1,14 +1,13 @@
 import "./App.css";
-import { Auth } from "./components/auth";
+import { Navbar } from "./components/navbar";
 import { Browsing  } from "./components/browsing";
-
 
 function App() {
   return (
     <div className="App">
-            <Auth />
+      <Navbar />
 
-            <Browsing />
+      <Browsing />
     </div>
   );
 }
