+++ src/components/signin.js
@@ -1,8 +1,9 @@
 import { auth, googleProvider } from '../config/firebase';
-import { createUserWithEmailAndPassword, signInWithPopup, signOut } from 'firebase/auth';
+import { createUserWithEmailAndPassword, signInWithPopup } from 'firebase/auth';
 import { useState } from 'react';
 
-export const Auth = () => {
+// Login component
+export const LoginMenu = () => {
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
 
@@ -22,14 +23,6 @@ export const Auth = () => {
         }
     };
 
-    const logout = async () => {
-        try{
-            await signOut(auth);
-        } catch (err) {
-            console.error(err);
-        }
-    };
-
     return (
         <div>
             <input 
@@ -44,8 +37,6 @@ export const Auth = () => {
             <button onClick={signIn}> Sign In</button>
 
             <button onClick={signInWithGoogle}> Sign In With Google</button>
-            
-            <button onClick={logout}> Logout </button>
         </div>
     );
-};
+}
